+++ R/dbplyr.R
@@ -23,13 +23,6 @@
 #'
 #' poolClose(pool)
 tbl.Pool <- function(src, from, ..., vars = NULL) {
-  con <- poolCheckout(src)
-  on.exit(poolReturn(con))
-
-  if (is.null(vars)) {
-    vars <- dplyr::db_query_fields(con, from)
-  }
-
   dbplyr::tbl_sql("Pool", dbplyr::src_dbi(src), from, ..., vars = vars)
 }
 
@@ -92,10 +85,13 @@ dbplyr_register_methods <- function() {
     dbplyr_s3_register("db_collect")
     dbplyr_s3_register("db_compute")
     dbplyr_s3_register("db_connection_describe")
+    dbplyr_s3_register("db_copy_to")
+    dbplyr_s3_register("db_col_types")
     dbplyr_s3_register("db_sql_render")
     dbplyr_s3_register("sql_translation")
     dbplyr_s3_register("sql_join_suffix")
     dbplyr_s3_register("sql_query_explain")
+    dbplyr_s3_register("sql_query_fields")
   })
 }